Make the store error seam structural via MappedDatabase
