+++ Sources/ClawData/Stores/Bot/AllowlistStoreGRDB.swift
@@ -3,14 +3,14 @@ import Foundation
 import GRDB
 
 public struct AllowlistStoreGRDB: AllowlistStore {
-  private let writer: any DatabaseWriter
+  private let database: MappedDatabase
 
   public init(writer: any DatabaseWriter) {
-    self.writer = writer
+    database = MappedDatabase(writer: writer)
   }
 
   public func seedAllowlist(userIds: [Int64]) throws {
-    try writer.writeMapping { db in
+    try database.writeMapping { db in
       for userId in userIds {
         try db.execute(
           sql: "INSERT OR IGNORE INTO allowlist(user_id, added_at) VALUES (?, ?)",
@@ -21,7 +21,7 @@ public struct AllowlistStoreGRDB: AllowlistStore {
   }
 
   public func allowlistContains(userId: Int64) throws -> Bool {
-    try writer.readMapping { db in
+    try database.readMapping { db in
       try Bool.fetchOne(
         db,
         sql: "SELECT EXISTS(SELECT 1 FROM allowlist WHERE user_id = ?)",
@@ -31,7 +31,7 @@ public struct AllowlistStoreGRDB: AllowlistStore {
   }
 
   public func allowlistCount() throws -> Int {
-    try writer.readMapping { db in
+    try database.readMapping { db in
       try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM allowlist") ?? 0
     }
   }
